+++ src/bfsetting/mod.rs
@@ -1,0 +1,7 @@
+//! Defines settings for the interpreter, compiler, and related tools.
+//!
+//! This module defines the setting struct to hold configuration options.
+
+mod setting;
+
+pub use setting::*;