+++ CHANGELOG.md
@@ -10,8 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Brainfuck interpreter options:
-    - Translating newline characters to value 0x10 (or vice versa (or both)).
+    - Translating newline characters to value 0x0A.
     - Configuring what value EOF is converted to: 0, -1 (for example, 0xFF in an 8-bit cell), or not writing any value (keeping the cell unchanged)
+    - Configuring what size each cell is: 8bit, 16bit, 32bit, 64bit or 128bit.
+        - Interpreter will read only one byte when reading input.
+        - Interpreter will write all bytes of cell when printing.
+    - Configuring which endian interprets the cell as character(s) when printing: Little-endian, Big-endian, or depending on the platform.
 
 ### Changed